Data Utility Class (Part 1)
Created enums for Project and Ticket Priorities, Ticket Statuses, and Ticket Types.

Created a Data Utility class to handle the connection to the database, regardless of whether it is local or on some platform such as Heroku or Railway. This class also will seed some default companies and other relevant information into the application.