Split Queries
Some of the data that we will query for the bug tracker can require multiple joined database queries that may have repeated data, and enabling query splitting will allow us to obtain the data we need without the unnecessary repeated data. This will help the performance of the application considerably.