Project Features - Add New Project POST Method
Built out the http post version of the Project controller's Create action.

In the process of testing this out, encountered Postgres Exception 23503 foreign key constraint violation. The ProjectPriorityId foreign key in the Project model was not nullable, so a value of 0 was assigned, which caused the error (values from 1 to 4 are expected for this). Made the FK nullable, and ran another migration and updated the database, which fixed the issue.