Project Features - Archive
Adjusted the Delete action and view to use "Archive" instead. Made some further adjustments to the action's code, and cleaned up the view. Finally, made some changes to the conversion of the image byte array to an actual image in the Index action, and copied that over to the Archive action.