Change to Project View - Details
Modified the layout of the history shown in the Project Activity section to match that shown in the Ticket Details.

The issue was that when a comment or attachment was added, the Previous Value and Current Value were showing with nothing displayed, implying that something was edited as opposed to being created.